server/requirements.txt
algoliasearch==1.17.0
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
Cython==0.29
Flask==1.0.2
google-api-core==1.5.1
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.0
matplotlib==3.0.1
nltk==3.3
numpy==1.15.3
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycocotools==2.0.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.0
rsa==4.0
six==1.11.0
torch==0.4.1
torchvision==0.2.1
urllib3==1.24.1
waitress==1.1.0
Werkzeug==0.14.1
